Дорабатывает task-comments

+++ project/apps/tasks/src/app/task-comment/task-comment.service.ts
@@ -1,18 +1,26 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable, UnauthorizedException } from '@nestjs/common';
 import { TaskCommentRepository } from './task-comment.repository';
 import { CreateCommentDto } from './dto/create-comment.dto';
 import { TaskCommentEntity } from './task-comment.entity';
 import { TaskCommentQuery } from './query/task-comment.query';
+import { JwtService } from '@nestjs/jwt';
+import { TaskCommentException } from './task-comment.constant';
 
 @Injectable()
 export class TaskCommentService {
   constructor(
     private readonly taskCommentRepository: TaskCommentRepository,
+    private readonly jwtService: JwtService
   ) {}
 
-  public async createComment(dto: CreateCommentDto, taskId: number) {
-    const commentEntity = new TaskCommentEntity(dto);
-    commentEntity.taskId = taskId;
+  public async createComment(dto: CreateCommentDto, taskId: number, token?: string) {
+    const user = this.jwtService.decode(token);
+
+    if (!user) {
+      throw new UnauthorizedException(TaskCommentException.Unauthorized);
+    }
+
+    const commentEntity = new TaskCommentEntity({...dto, userId: user['sub'], taskId});
 
     return this.taskCommentRepository
       .create(commentEntity);
@@ -22,7 +30,18 @@ export class TaskCommentService {
     return this.taskCommentRepository.findById(id);
   }
 
-  public async deleteComment(id: number) {
+  public async deleteComment(id: number, token?: string) {
+    const user = this.jwtService.decode(token);
+    const comment = await this.taskCommentRepository.findById(id);
+
+    if (!user) {
+      throw new UnauthorizedException(TaskCommentException.Unauthorized);
+    }
+
+    if (user['sub'] !== comment.userId) {
+      throw new ForbiddenException(TaskCommentException.Forbidden);
+    }
+
     this.taskCommentRepository.destroy(id);
   }
